refactor: put resource and document types into a renamed file
