+++ packages/host/app/components/prerendered-card-search.gts
@@ -21,7 +21,7 @@ import {
 import {
   PrerenderedCardCollectionDocument,
   isPrerenderedCardCollectionDocument,
-} from '@cardstack/runtime-common/card-document';
+} from '@cardstack/runtime-common/document-types';
 
 import { type Format } from 'https://cardstack.com/base/card-api';
