+++ packages/realm-server/tests/server-endpoints-test.ts
@@ -44,7 +44,7 @@ import '@cardstack/runtime-common/helpers/code-equality-assertion';
 import { RealmServer } from '../server';
 import { MatrixClient } from '@cardstack/runtime-common/matrix-client';
 import jwt from 'jsonwebtoken';
-import { type CardCollectionDocument } from '@cardstack/runtime-common/card-document';
+import { type CardCollectionDocument } from '@cardstack/runtime-common/document-types';
 import { type PgAdapter } from '@cardstack/postgres';
 import {
   getUserByMatrixUserId,
