+++ packages/runtime-common/resource-types.ts
@@ -1,8 +1,21 @@
-import { type CodeRef, isCodeRef } from './code-ref';
 import { RealmInfo } from './realm';
-import { QueryResultsMeta, PrerenderedCard } from './index-query-engine';
-import { type CardTypeSummary } from './index-structure';
+import { type CodeRef, isCodeRef } from './code-ref';
 
+// resource
+export type Resource = ModuleResource | CardResource | PrerenderedCardResource;
+export type ResourceMeta = ModuleMeta | Meta;
+
+//modules
+export type ModuleMeta = {};
+
+export interface ModuleResource {
+  id?: string;
+  type: 'source';
+  attributes?: { content: string };
+  meta: ModuleMeta;
+}
+
+//cards
 export type Saved = string;
 export type Unsaved = string | undefined;
 export interface Meta {
@@ -56,25 +69,29 @@ export interface CardResource<Identity extends Unsaved = Saved> {
     self?: string;
   };
 }
-export interface SingleCardDocument<Identity extends Unsaved = Saved> {
-  data: CardResource<Identity>;
-  included?: CardResource<Saved>[];
-}
-export interface CardCollectionDocument<Identity extends Unsaved = Saved> {
-  data: CardResource<Identity>[];
-  included?: CardResource<Saved>[];
-}
 
-export interface PrerenderedCardCollectionDocument {
-  data: PrerenderedCardResource[];
-  meta: QueryResultsMeta & {
-    scopedCssUrls?: string[];
-    realmInfo?: RealmInfo;
+//prerendered cards
+export interface PrerenderedCardResource {
+  id: string;
+  type: 'prerendered-card';
+  attributes: {
+    html: string;
+    isError?: true;
+  };
+  relationships: {
+    'prerendered-card-css': {
+      data: { id: string }[];
+    };
+  };
+  meta: Meta & {};
+  links?: {
+    self?: string;
   };
 }
 
-export type CardDocument = SingleCardDocument | CardCollectionDocument;
+//validation - modules
 
+//validation - cards
 export function isCardResource(resource: any): resource is CardResource {
   if (typeof resource !== 'object' || resource == null) {
     return false;
@@ -122,7 +139,6 @@ export function isCardResource(resource: any): resource is CardResource {
   let { adoptsFrom } = meta;
   return isCodeRef(adoptsFrom);
 }
-
 export function isCardFields(fields: any): fields is CardFields {
   if (typeof fields !== 'object') {
     return false;
@@ -217,95 +233,7 @@ export function isRelationship(
   return true;
 }
 
-export function isCardDocument(doc: any): doc is CardDocument {
-  return isSingleCardDocument(doc) || isCardCollectionDocument(doc);
-}
-
-export function isCardDocumentString(maybeJsonString: string) {
-  try {
-    let doc = JSON.parse(maybeJsonString);
-    return isSingleCardDocument(doc) || isCardCollectionDocument(doc);
-  } catch (err) {
-    return false;
-  }
-}
-
-export function isSingleCardDocument(doc: any): doc is SingleCardDocument {
-  if (typeof doc !== 'object' || doc == null) {
-    return false;
-  }
-  if (!('data' in doc)) {
-    return false;
-  }
-  let { data } = doc;
-  if (Array.isArray(data)) {
-    return false;
-  }
-  if ('included' in doc) {
-    let { included } = doc;
-    if (!isIncluded(included)) {
-      return false;
-    }
-  }
-  return isCardResource(data);
-}
-
-export function isCardCollectionDocument(
-  doc: any,
-): doc is CardCollectionDocument {
-  if (typeof doc !== 'object' || doc == null) {
-    return false;
-  }
-  if (!('data' in doc)) {
-    return false;
-  }
-  let { data } = doc;
-  if (!Array.isArray(data)) {
-    return false;
-  }
-  if ('included' in doc) {
-    let { included } = doc;
-    if (!isIncluded(included)) {
-      return false;
-    }
-  }
-  return data.every((resource) => isCardResource(resource));
-}
-
-export function isPrerenderedCardCollectionDocument(
-  doc: any,
-): doc is PrerenderedCardCollectionDocument {
-  if (typeof doc !== 'object' || doc == null) {
-    return false;
-  }
-  if (!('data' in doc) || !('meta' in doc)) {
-    return false;
-  }
-  let { data } = doc;
-  if (!Array.isArray(data)) {
-    return false;
-  }
-  return data.every((resource) => isPrerenderedCardResource(resource));
-}
-
-export interface PrerenderedCardResource {
-  id: string;
-  type: 'prerendered-card';
-  attributes: {
-    html: string;
-    isError?: true;
-  };
-  relationships: {
-    'prerendered-card-css': {
-      data: { id: string }[];
-    };
-  };
-  meta: Meta & {};
-  links?: {
-    self?: string;
-  };
-}
-
+//validation - prerendered cards
 export function isPrerenderedCardResource(
   resource: any,
 ): resource is PrerenderedCardResource {
@@ -322,64 +250,4 @@ export function isPrerenderedCardResource(
     return false;
   }
   return true;
-}
-
-function isIncluded(included: any): included is CardResource<Saved>[] {
-  if (!Array.isArray(included)) {
-    return false;
-  }
-  for (let resource of included) {
-    if (typeof resource !== 'object' || !resource) {
-      return false;
-    }
-    if (
-      (!('id' in resource) || typeof resource.id !== 'string') &&
-      (!('lid' in resource) || typeof resource.lid !== 'string')
-    ) {
-      return false;
-    }
-    if (!isCardResource(resource)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-export function transformResultsToPrerenderedCardsDoc(results: {
-  prerenderedCards: PrerenderedCard[];
-  scopedCssUrls: string[];
-  meta: QueryResultsMeta & { scopedCssUrls?: string[]; realmInfo?: RealmInfo };
-}) {
-  let { prerenderedCards, scopedCssUrls, meta } = results;
-
-  let data = prerenderedCards.map((card) => ({
-    type: 'prerendered-card',
-    id: card.url,
-    attributes: {
-      html: card.html,
-      usedRenderType: card.usedRenderType,
-      ...(card.isError ? { isError: true as const } : {}),
-    },
-  }));
-
-  meta.scopedCssUrls = scopedCssUrls;
-
-  return {
-    data,
-    meta,
-  };
-}
-
-export function makeCardTypeSummaryDoc(summaries: CardTypeSummary[]) {
-  let data = summaries.map((summary) => ({
-    type: 'card-type-summary',
-    id: summary.code_ref,
-    attributes: {
-      displayName: summary.display_name,
-      total: summary.total,
-      iconHTML: summary.icon_html,
-    },
-  }));
-
-  return { data };
 }
